First version with own graph.

+++ donkeycar/parts/tflite.py
@@ -1,5 +1,6 @@
 import os
 import tensorflow as tf
+import numpy as np
 
 from donkeycar.parts.keras import KerasPilot
 
@@ -55,7 +56,7 @@ class TFLitePilot(KerasPilot):
         self.input_shape = self.input_details[0]['shape']
     
     def inference(self, img_arr, other_arr):
-        input_data = img_arr.reshape(self.input_shape)
+        input_data = img_arr.reshape(self.input_shape).astype(np.float32)
         self.interpreter.set_tensor(self.input_details[0]['index'], input_data)
         self.interpreter.invoke()
